requirements.txt
streamlit
plost
holoviews
hvplot
python-dotenv
pandas
numpy
pathlib
matplotlib
geoviews 
cartopy
